ad 102 fabionacciu fix3

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,16 +9,16 @@
 int main(void)
 {
 	int i;
-	unsigned int a = 0;
-	unsigned int b = 1;
-	unsigned int res = 0;
+	int a = 0;
+	int b = 1;
+	unsigned long long int res = 0;
 
 	for (i = 1; i <= 50; i++)
 	{
 		res = a + b;
 		a = b;
 		b = res;
-		printf("%u, ", res);
+		printf("%llu, ", res);
 
 	}
 	printf("\n");
